// src/app/api/fn/mosques/get-mosques.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';


export interface GetMosques$Params {
  Name?: string | null;
  City?: string | null;
  Country?: string | null;
  Governorate?: string | null;
  District?: string | null;
  IsActive?: boolean | null;
  IsVerified?: boolean | null;
  PageNumber?: number | null;
  PageSize?: number | null;
}

export function getMosques(http: HttpClient, rootUrl: string, params?: GetMosques$Params, context?: HttpContext): Observable<StrictHttpResponse<void>> {
  const rb = new RequestBuilder(rootUrl, getMosques.PATH, 'get');
  if (params) {
    rb.query('Name', params.Name, {});
    rb.query('City', params.City, {});
    rb.query('Country', params.Country, {});
    rb.query('Governorate', params.Governorate, {});
    rb.query('District', params.District, {});
    rb.query('IsActive', params.IsActive, {});
    rb.query('IsVerified', params.IsVerified, {});
    rb.query('PageNumber', params.PageNumber, {});
    rb.query('PageSize', params.PageSize, {});
  }

  return http.request(
    rb.build({ responseType: 'text', accept: '*/*', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return (r as HttpResponse<any>).clone({ body: undefined }) as StrictHttpResponse<void>;
    })
  );
}

getMosques.PATH = '/api/Mosques';